requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
pyyaml==5.4.1